# data/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Athena CTF",
    "description": "# Oxide\n\nOxide is the api server of [Athena CTF](https://github.com/athena-ctf/athena-ctf) platform.\n\n## Usage\n\n```\noxide run /data/config.toml\n```\n",
    "contact": {
      "name": "AthenaCTF Team",
      "email": "athena.help@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://www.mit.edu/~amini/LICENSE.md"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/achivement/{id}": {
      "get": {
        "tags": [
          "achievement"
        ],
        "summary": "Retrieve achievement by id",
        "operationId": "retrieve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved achievement by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AchievementModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No achievement found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/challenge": {
      "get": {
        "tags": [
          "challenge"
        ],
        "summary": "List challenges",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Listed challenges successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChallengeModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "challenge"
        ],
        "summary": "Create challenge",
        "operationId": "create",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreateChallengeSchema"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created challenge successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/challenge/{id}/relations": {
      "get": {
        "tags": [
          "challenge"
        ],
        "summary": "Retrieve challenge relations by id",
        "operationId": "retrieve_relations_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved challenge relations by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeRelations"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No challenge found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
    "/challenge/{id}/solves": {
      "get": {
        "tags": [
          "challenge"
        ],
        "summary": "Calculate challenge solves by id",
        "operationId": "calculate_solves_by_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calculates challenge solves by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChallengeSolves"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No challenge found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/instance/{id}": {
      "get": {
        "tags": [
          "instance"
        ],
        "summary": "Retrieve instance by id",
        "operationId": "retrieve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved instance by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No instance found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "instance"
        ],
        "summary": "Delete instance by id",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted instance by id successfully"
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No instance found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "instance"
        ],
        "summary": "Update instance by id",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InstanceDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated instance by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstanceModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters/request body format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No instance found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
    "/invite/{id}": {
      "get": {
        "tags": [
          "invite"
        ],
        "summary": "Retrieve invite by id",
        "operationId": "retrieve",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved invite by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No invite found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "invite"
        ],
        "summary": "Delete invite by id",
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted invite by id successfully"
          },
          "400": {
            "description": "Invalid parameters format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No invite found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "invite"
        ],
        "summary": "Update invite by id",
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Id of entity",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InviteDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated invite by id successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InviteModel"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters/request body format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "401": {
            "description": "Action is permissible after login",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "403": {
            "description": "User does not have sufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "404": {
            "description": "No invite found with specified id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AchievementModel": {
        "type": "object",
        "required": [
          "id",
          "value",
          "date_created",
          "prize",
          "challenge_id"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          },
          "date_created": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "player_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "prize": {
            "type": "integer",
            "format": "int32"
          },
          "team_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "value": {
            "type": "string"
          }
        }
      },
...
      "ChallengeModel": {
        "type": "object",
        "required": [
          "id",
          "title",
          "description",
          "min_points",
          "max_points",
          "points",
          "difficulty",
          "flag_type",
          "status",
          "author_name"
        ],
        "properties": {
          "author_name": {
            "type": "string"
          },
          "container_details": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ContainerMeta"
              }
            ],
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyEnum"
          },
          "flag_type": {
            "$ref": "#/components/schemas/FlagTypeEnum"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "max_points": {
            "type": "integer",
            "format": "int32"
          },
          "min_points": {
            "type": "integer",
            "format": "int32"
          },
          "points": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/ChallengeStatusEnum"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "ChallengeRelations": {
        "type": "object",
        "required": [
          "challenge",
          "files",
          "hints",
          "tags",
          "solves"
        ],
        "properties": {
          "challenge": {
            "$ref": "#/components/schemas/ChallengeModel"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileModel"
            }
          },
          "hints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChallengeHint"
            }
          },
          "solves": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagModel"
            }
          }
        }
      },
      "ChallengeSolves": {
        "type": "object",
        "required": [
          "solves"
        ],
        "properties": {
          "solves": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
...
      "CreateChallengeSchema": {
        "type": "object",
        "required": [
          "container_details",
          "author_name",
          "description",
          "difficulty",
          "max_points",
          "min_points",
          "status",
          "title",
          "tags",
          "flag_type"
        ],
        "properties": {
          "author_name": {
            "type": "string"
          },
          "container_details": {
            "type": "string",
            "format": "binary"
          },
          "description": {
            "type": "string"
          },
          "difficulty": {
            "$ref": "#/components/schemas/DifficultyEnum"
          },
          "flag_type": {
            "$ref": "#/components/schemas/FlagTypeEnum"
          },
          "max_points": {
            "type": "integer",
            "format": "int32"
          },
          "min_points": {
            "type": "integer",
            "format": "int32"
          },
          "single_instance": {
            "type": "boolean",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ChallengeStatusEnum"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ErrorModel": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "InstanceDetails": {
        "type": "object",
        "required": [
          "challenge_id",
          "player_id"
        ],
        "properties": {
          "challenge_id": {
            "type": "string",
            "format": "uuid"
          },
          "player_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}